backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
motor==3.3.2
pymongo==4.5.0
PyYAML==6.0.1
websockets==12.0
httpx==0.27.0
